Refactor: ClubRepository 관련 중복 제거
내용:
충돌이 안나기위해 기존에
ClubRepository  ClubRepository2
-> 하나로 ClubRepository 로 통합 (club 패키지에)
