+++ backend/src/main/java/com/example/smu_club/club/service/ClubService.java
@@ -1,16 +1,16 @@
 package com.example.smu_club.club.service;
 
-import com.example.smu_club.auth.repository.MemberRepository;
 import com.example.smu_club.club.dto.ApplicationFormResponseDto;
 import com.example.smu_club.club.dto.ClubResponseDto;
 import com.example.smu_club.club.dto.ClubsResponseDto;
-import com.example.smu_club.club.repository.ClubRepository;
 import com.example.smu_club.domain.Club;
 import com.example.smu_club.domain.Member;
 import com.example.smu_club.domain.Question;
 import com.example.smu_club.exception.custom.ClubNotFoundException;
 import com.example.smu_club.exception.custom.MemberNotFoundException;
+import com.example.smu_club.member.repository.MemberRepository;
 import com.example.smu_club.question.dto.QuestionResponse;
+import com.example.smu_club.question.repository.ClubRepository;
 import com.example.smu_club.question.repository.QuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -63,6 +63,7 @@ public class ClubService {
         List<QuestionResponse> clubQuestionListResponse =
                 questionList.stream().map
                         (qr -> new QuestionResponse(
+                                qr.getId(),
                                 qr.getContent(),
                                 qr.getOrderNum()
                         )).collect(toList());
