+++ backend/src/main/java/com/example/smu_club/question/repository/ClubRepository.java
@@ -1,9 +1,10 @@
-package com.example.smu_club.club.repository;
+package com.example.smu_club.question.repository;
 
 import com.example.smu_club.domain.Club;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface ClubRepository extends JpaRepository<Club, Long> {
+
+
 }
+
