--- backend/src/main/java/com/example/smu_club/question/repository/ClubRepository2.java
@@ -1,9 +1,0 @@
-package com.example.smu_club.question.repository;
-
-import com.example.smu_club.domain.Club;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ClubRepository2 extends JpaRepository<Club, Long> {
-
-
-}
